adds messaging mapper impl

+++ infrastructure/kafka/src/lib.rs
@@ -13,19 +13,20 @@ pub mod model {
 
         pub mod payment_request {
             use apache_avro::AvroSchema;
-            #[derive(apache_avro::AvroSchema)]
+            use derive_builder::Builder;
+            #[derive(apache_avro::AvroSchema, Clone)]
             pub enum PaymentOrderStatus {
                 PENDING,
                 CANCELLED,
             }
 
-            #[derive(AvroSchema)]
+            #[derive(AvroSchema, Builder)]
             pub struct PaymentRequest {
                 pub id: uuid::Uuid,
                 pub saga_id: uuid::Uuid,
                 pub customer_id: uuid::Uuid,
                 pub order_id: uuid::Uuid,
-                pub price: f64,
+                pub price: i64,
                 pub created_at: i64,
                 pub payment_order_status: PaymentOrderStatus,
             }
